ai-services/fall-detection/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
librosa==0.10.1
soundfile==0.12.1
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2
tensorflow==2.15.0
pyaudio==0.2.11
webrtcvad==2.0.10
requests==2.31.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
asyncio-mqtt==0.13.0
